blog-api/requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.0
Django==5.0.7
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
sqlparse==0.5.1
tzdata==2024.1
